requirements.txt
torch>=1.1.0
torchvision
numpy
scipy
Pillow
scikit-image
opencv-contrib-python
tqdm
